Move metadata content and fix Storybook styles
