+++ node-src/ui/html/metadata.html.ts
@@ -11,7 +11,6 @@ export default ({ announcedBuild, build }: Context, files: FileDesc[]) => `<!DOC
     <title>Build ${announcedBuild.number} metadata files</title>
     <style>
       body {
-        -webkit-font-smoothing: antialiased;
         font-family: 'Nunito Sans', sans-serif;
         font-size: 14px;
         line-height: 1;
